Complete integration with the workflow of QiBuild
Few pieces were missing for the qibuildization to be complete, mainly
because QiBuild requires the project to be usable as a dependency
directly from the build directory. To sum up, all output should land in
the same directory tree so that they can be found back by Soprano's,
Qt's and QiBuild's mechanisms. The look-up by the plug-in manager
also had to be adapted accordingly.

Change-Id: I884608f7b5e5337e55cb36dd71dd1db1d1686753