Restore original CMake code
Policy has changed about how to deal with this repository.
In the end, it will not be used in the worktree directly,
and will require to be built and packaged separately from
the rest of the code, so that it is used in a toolchain.

This change reverts the integrality of the changes in the
CMake code, and removes any occurrence of libQi, that was
used to find the libraries, plugins, or to produce logs.
Boost is not used anymore for function objects, instead
the STL is preferred.

But this change keeps the change from raptor to raptor2,
the qiproject.xml, and the logs added to debug the look-up
of plugins.